feat: move default game scene to level1scene
