+++ rsclone/src/js/main.js
@@ -10,7 +10,7 @@ class Main {
   constructor() {
     const cookieVersion = 0;
     const settings = JSON.parse(localStorage.getItem('rsc-game-settings')) || {
-      cookieVersion
+      cookieVersion,
     };
     const savedVersion = settings.cookieVersion;
     if (savedVersion !== cookieVersion) localStorage.clear();
